pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skipalign"
version = "0.1.0"
description = "Alignment-free conserved region discovery and RT-qPCR primer-probe design for highly divergent viral genera"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Key-man-fromArchive" },
]
keywords = ["bioinformatics", "k-mer", "RT-qPCR", "primer-design", "alignment-free", "orthoflavivirus"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9",
    "biopython>=1.83",
    "numpy>=1.24",
    "scipy>=1.10",
    "primer3-py>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
fast = ["jellyfish"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.scripts]
skipalign = "skipalign.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/skipalign"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
